code/web/requirements.txt
aioredis==1.2.0
amqp==2.5.0
asgi-redis==1.4.3
asgiref==2.3.0
astroid==2.1.0
async-timeout==2.0.1
asyncio==3.4.3
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
beautifulsoup4==4.6.3
billiard==3.6.0.0
cachetools==3.0.0
celery==4.3.0
certifi==2018.10.15
channels==2.1.6
channels-redis==2.3.2
chardet==3.0.4
cloudinary==1.15.0
constantly==15.1.0
coverage==4.5.3
daphne==2.2.4
DateTime==4.3
defusedxml==0.5.0
dill==0.3.1.1
dj-database-url==0.5.0
Django==2.1.3
django-cloudinary-storage==0.2.3
django-cors-headers==3.2.1
django-debug-toolbar==1.10.1
django-file-resubmit==0.5.2
django-filter==2.2.0
django-heroku==0.3.1
django-jet==1.0.8
django-nose==1.4.6
django-pure-pagination==0.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
google-api-python-client==1.7.5
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.7.1
hiredis==0.3.1
httplib2==0.12.0
hyperlink==18.0.0
idna==2.7
isort==4.3.4
klepto==0.1.8
kombu==4.6.1
lazy-object-proxy==1.3.1
Markdown==3.2.1
mccabe==0.6.1
mock==2.0.0
msgpack==0.5.6
msgpack-python==0.5.6
mysql-connector-python==8.0.16
mysqlclient==1.3.13
nose==1.3.7
numpy==1.17.0
oauth2client==4.1.3
oauthlib==2.1.0
packaging==19.0
pbr==5.1.1
Pillow==5.3.0
pip-review==1.0
pox==0.2.7
protobuf==3.11.3
psycopg2==2.7.4
psycopg2-binary==2.8.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyHamcrest==1.9.0
PyJWT==1.6.4
pykakasi==1.2
pylint==2.2.2
pyparsing==2.3.1
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.7
redis==3.0.1
regex==2019.3.12
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
selenium==3.141.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
SQLAlchemy==1.2.14
sqlparse==0.2.4
stripe==2.20.3
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.3.0
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.24.1
user-agents==1.1.0
uWSGI==2.0.17.1
vine==1.3.0
whitenoise==5.0.1
wrapt==1.11.1
zope.interface==4.6.0
